Moves Player class into its own file
